+++ test/matchRoute.spec.js
@@ -74,7 +74,7 @@ describe('matchRoute(route, baseUrl, path)', () => {
 
   it('should match 2 routes (1)', () => {
     const route = {
-      path: '/',
+      path: '',
       children: [
         {
           path: '/a',
@@ -85,7 +85,7 @@ describe('matchRoute(route, baseUrl, path)', () => {
     expect(result).to.have.lengthOf(2);
     expect(result[0]).to.have.property('baseUrl', '');
     expect(result[0]).to.have.property('path', '/');
-    expect(result[0]).to.have.nested.property('route.path', '/');
+    expect(result[0]).to.have.nested.property('route.path', '');
     expect(result[1]).to.have.property('baseUrl', '');
     expect(result[1]).to.have.property('path', '/a');
     expect(result[1]).to.have.nested.property('route.path', '/a');